Renamed creds.js to config.js
