+++ .gitignore
@@ -1,5 +1,3 @@
 node_modules/
 yarn.lock
-creds.js
 download/
-creds.js
