+++ src/app.js
@@ -1,6 +1,6 @@
 const showBanner = require("node-banner");
-const creds = require("../creds");
-const { processData } = require("./processData");
+const config = require("../config");
+// const { processData } = require("./processData");
 const { removeAllFilesInDirectory } = require("./removeAllFiles");
 const { scrape } = require("./scrape");
 
@@ -9,7 +9,7 @@ const main = async () => {
     "FIXR  SCRAPER",
     "Queen's Computing Society, Queen's University Belfast\n 2022 - James McFarland."
   );
-  await removeAllFilesInDirectory(creds.downloadDir);
+  await removeAllFilesInDirectory(config.downloadDir);
   await scrape(false);
   // await processData();
   return
