+++ src/index.js
@@ -1,5 +1,5 @@
 const schedule = require("node-schedule");
-const creds = require("../creds");
+const config = require("../config");
 const { main } = require("./app");
 
 process.on("SIGINT", function () {
@@ -7,7 +7,7 @@ process.on("SIGINT", function () {
 });
 
 //Run at minute 0 of every hour
-schedule.scheduleJob(`0 */${creds.runEvery} * * *`, () => {
+schedule.scheduleJob(`0 */${config.runEvery} * * *`, () => {
   console.log("Starting run at " + new Date());
   main();
   console.log("Run complete")
