+++ src/processData.js
@@ -2,12 +2,12 @@ const { parse } = require("csv-parse");
 require("console-stamp")(console, { format: ":date(HH:MM:ss.l)" });
 const fs = require("fs");
 const path = require("path");
-const creds = require("../creds");
+const config = require("../config");
 const { googleSheets } = require("./sheets");
 
 const processData = async () => {
   console.log("[PD] Reading CSV...");
-  fs.readdir(creds.downloadDir, (err, files) => {
+  fs.readdir(config.downloadDir, (err, files) => {
     if (err) {
       console.log("[PD] Error processing files (READDIR)");
       return;
@@ -26,7 +26,7 @@ const processData = async () => {
         googleSheets(records);
       });
 
-      fs.createReadStream(path.join(creds.downloadDir, file)).pipe(parser);
+      fs.createReadStream(path.join(config.downloadDir, file)).pipe(parser);
     }
   });
 };
